+++ packages/deposit-screener/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- fix server cli to also take `--dummy-screener-delay`
 - add dummy magic rejection value for screener to reject some deposits (test purposes)
 - put back redis utils
 - make logging more consistent